sensorhub: brcm: workaround for -Wenum-compare
../drivers/sensorhub/brcm/ssp_dev.c:774:21: error: comparison of different enumeration types ('enum panel_notifier_event_panel_power_state' and 'enum panel_notifier_event_copr_state') [-Werror,-Wenum-compare]
  774 |                 if(evt_data.state != PANEL_EVENT_COPR_ENABLED && evt_data.state != PANEL_EVENT_COPR_DISABLED){
      |                    ~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~
../drivers/sensorhub/brcm/ssp_dev.c:774:67: error: comparison of different enumeration types ('enum panel_notifier_event_panel_power_state' and 'enum panel_notifier_event_copr_state') [-Werror,-Wenum-compare]
  774 |                 if(evt_data.state != PANEL_EVENT_COPR_ENABLED && evt_data.state != PANEL_EVENT_COPR_DISABLED){
      |                                                                  ~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~
2 errors generated.

+++ drivers/sensorhub/brcm/ssp_dev.c
@@ -771,7 +771,7 @@ static int panel_notifier_callback(struct notifier_block *self, unsigned long ev
 	} else if (event == PANEL_EVENT_COPR_STATE_CHANGED) {
 		struct panel_power_state_event_data evt_data = *((struct panel_power_state_event_data *)data);		
 		pr_info("[SSP] %s PANEL_EVENT_COPR_STATE_CHANGED, event(%d)\n", __func__, evt_data.state);
-		if(evt_data.state != PANEL_EVENT_COPR_ENABLED && evt_data.state != PANEL_EVENT_COPR_DISABLED){
+		if((int)evt_data.state != (int)PANEL_EVENT_COPR_ENABLED && (int)evt_data.state != (int)PANEL_EVENT_COPR_DISABLED){
 			pr_info("[SSP] %s PANEL_EVENT_COPR_STATE_CHANGED, event errno(%d)\n", __func__, evt_data.state);
 		} else {
 			ssp_data_info->uLastAPState = evt_data.state == 0 ? MSG2SSP_AP_STATUS_SLEEP : MSG2SSP_AP_STATUS_WAKEUP;
